src/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
distlib==0.3.3
Django==4.0
django-cors-headers==3.10.1
django-debug-toolbar==3.2.4
django-dotenv==1.4.2
django-restframework==0.0.1
djangorestframework==3.13.1
filelock==3.4.0
gunicorn==20.1.0
idna==3.3
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
Pillow==8.4.0
pipenv==2021.11.23
platformdirs==2.4.0
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.10.0
virtualenv-clone==0.5.7
wincertstore==0.2
